User is now able to login if not in database, but has temporary authority of user.
On successful login a new user is created in the home controller.

Here if the user is not in the session then we add the user with default values to DB.

Then we add UserRole for said user to the DB with default roleTypeID =3 for chairman

There is code for when we have attributes but for now the else if statement does much the same thing
but with the default values stated instead of the details we would get from CAS's attribute release.

I have changed the proposal controller to get the UserID from the User stored in Session.
Look at the last method for how this is done.